requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database - PostgreSQL for production
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Configuration and validation
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==3.17.1

# AI and embeddings - stable versions for production
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.6.1
sentence-transformers==2.2.2

# Vector database
chromadb==0.4.18

# Data processing
numpy==1.24.3
pandas==2.0.3

# Logging and monitoring
structlog==23.1.0
prometheus-client==0.19.0

# Production deployment
gunicorn==21.2.0

# Optional: Redis for caching (if using)
redis==5.0.1
aioredis==2.0.1